+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.17
 - :rocket: added _I force click_ step
+- :beetle: added support of memory values in _I click '$value' text in 'collection' collection_
 
 ## 0.0.16
 - :beetle: fixed optional params templates in wait steps